6-12-23

+++ app/Http/Controllers/ProductController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Models\Product;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class ProductController extends Controller
 {
@@ -13,8 +14,10 @@ class ProductController extends Controller
     public function index()
     {
         $products = Product::all();
+        $user = Auth::user();
 
-        return view('products.index', compact('products'));
+        // return view('products.index', compact('products'));
+        return view('latihan.manajemen_product', ["active" => "manajemen_product"], compact('products', 'user'));
     }
 
     /**
@@ -23,7 +26,11 @@ class ProductController extends Controller
     public function create()
     {
         //
-        return view('products.create');
+        $user = Auth::user();
+        // return view('products.create', compact('user'));
+        // return view('products.create', ['active' => 'manajemen_product'], compact('users'));
+
+        return view('latihan.create_product', ['active' => 'manajemen_product'], compact('user'));
     }
 
     /**
@@ -58,7 +65,9 @@ class ProductController extends Controller
      */
     public function edit(Product $product)
     {
-        return view('products.edit', compact('product'));
+        $user = Auth::user();
+        // return view('products.edit', compact('product'));
+        return view('latihan.edit_product', ["active" => "manajemen_product"], compact('product', 'user'));
     }
 
     /**
@@ -83,12 +92,26 @@ class ProductController extends Controller
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(Product $product)
+    public function destroy($id)
     {
         //
+        // $product->delete();
+
+        // return redirect()->route('products.index')
+        //     ->with('success', 'Product deleted successfully');
+
+        $product = Product::findOrFail($id);
+
+        // Cek apakah ada pesanan terkait dengan produk
+        if ($product->orders()->exists()) {
+            // Tampilkan peringatan SweetAlert jika produk memiliki pesanan terkait
+            return back()->with('error', 'Produk memiliki pesanan terkait. Anda tidak dapat menghapus produk ini.');
+        }
+
+        // Lanjutkan penghapusan jika tidak ada pesanan terkait
         $product->delete();
 
-        return redirect()->route('products.index')
-            ->with('success', 'Product deleted successfully');
+        // Redirect ke halaman yang sesuai dan berikan pesan sukses
+        return redirect()->route('products.index')->with('success', 'Produk berhasil dihapus.');
     }
 }
